Add DataSink + fix DataSource template file type
